Create ExchangeService and DiceRollService. Reorganize Game. Move related methods of Game to ExchangeService and DiceRollService accordingly.
